containerize product price and description job
